backend/requirements.txt
fastapi
uvicorn
hashids
sqlalchemy
psycopg2-binary
python-dotenv
pydantic[email]
redis
